+++ packages/multichain-account-service/src/types.ts
@@ -11,6 +11,8 @@ import type {
   AccountsControllerGetAccountByAddressAction,
   AccountsControllerListMultichainAccountsAction,
 } from '@metamask/accounts-controller';
+import type { TraceCallback } from '@metamask/controller-utils';
+import type { ErrorReportingServiceCaptureExceptionAction } from '@metamask/error-reporting-service';
 import type { KeyringAccount } from '@metamask/keyring-api';
 import type {
   KeyringControllerAddNewKeyringAction,
@@ -81,6 +83,11 @@ export type MultichainAccountServiceCreateMultichainAccountWalletAction = {
   handler: MultichainAccountService['createMultichainAccountWallet'];
 };
 
+export type MultichainAccountServiceResyncAccountsAction = {
+  type: `${typeof serviceName}:resyncAccounts`;
+  handler: MultichainAccountService['resyncAccounts'];
+};
+
 /**
  * All actions that {@link MultichainAccountService} registers so that other
  * modules can call them.
@@ -95,7 +102,8 @@ export type MultichainAccountServiceActions =
   | MultichainAccountServiceSetBasicFunctionalityAction
   | MultichainAccountServiceAlignWalletAction
   | MultichainAccountServiceAlignWalletsAction
-  | MultichainAccountServiceCreateMultichainAccountWalletAction;
+  | MultichainAccountServiceCreateMultichainAccountWalletAction
+  | MultichainAccountServiceResyncAccountsAction;
 
 export type MultichainAccountServiceMultichainAccountGroupCreatedEvent = {
   type: `${typeof serviceName}:multichainAccountGroupCreated`;
@@ -135,7 +143,8 @@ type AllowedActions =
   | KeyringControllerGetKeyringsByTypeAction
   | KeyringControllerAddNewKeyringAction
   | NetworkControllerGetNetworkClientByIdAction
-  | NetworkControllerFindNetworkClientIdByChainIdAction;
+  | NetworkControllerFindNetworkClientIdByChainIdAction
+  | ErrorReportingServiceCaptureExceptionAction;
 
 /**
  * All events published by other modules that {@link MultichainAccountService}
@@ -155,3 +164,7 @@ export type MultichainAccountServiceMessenger = Messenger<
   MultichainAccountServiceActions | AllowedActions,
   MultichainAccountServiceEvents | AllowedEvents
 >;
+
+export type MultichainAccountServiceConfig = {
+  trace?: TraceCallback;
+};