+++ packages/notification-services-controller/src/NotificationServicesController/processors/process-snap-notifications.test.ts
@@ -1,17 +1,19 @@
 import { processSnapNotification } from './process-snap-notifications';
+import type { INotification } from '..';
 import { TRIGGER_TYPES } from '../constants';
 import { createMockSnapNotification } from '../mocks';
 
 describe('process-snap-notifications - processSnapNotification()', () => {
   it('processes a Raw Snap Notification to a shared Notification Type', () => {
     const rawNotification = createMockSnapNotification();
-    const result = processSnapNotification(rawNotification);
+    const result = processSnapNotification(rawNotification) as Extract<
+      INotification,
+      { type: TRIGGER_TYPES.SNAP }
+    >;
 
     expect(result.type).toBe(TRIGGER_TYPES.SNAP);
     expect(result.isRead).toBe(false);
     expect(result.data).toBeDefined();
-    // @ts-expect-error readDate property is guaranteed to exist
-    // as we're dealing with a snap notification
     expect(result.readDate).toBeNull();
   });
 });