+++ packages/subscription-controller/src/index.ts
@@ -60,6 +60,12 @@ export type {
   UpdatePaymentMethodCardResponse,
   CachedLastSelectedPaymentMethod,
   SubmitSponsorshipIntentsMethodParams,
+  Cohort,
+  CohortName,
+  BalanceCategory,
+  AssignCohortRequest,
+  GetSubscriptionsEligibilitiesRequest,
+  ModalType,
 } from './types';
 export {
   CRYPTO_PAYMENT_METHOD_ERRORS,
@@ -68,6 +74,9 @@ export {
   RECURRING_INTERVALS,
   PAYMENT_TYPES,
   SubscriptionUserEvent,
+  COHORT_NAMES,
+  BALANCE_CATEGORIES,
+  MODAL_TYPE,
 } from './types';
 export { SubscriptionServiceError } from './errors';
 export { Env, SubscriptionControllerErrorMessage } from './constants';