+++ packages/transaction-pay-controller/src/strategy/relay/relay-quotes.ts
@@ -1,25 +1,33 @@
-import { successfulFetch, toHex } from '@metamask/controller-utils';
+import { Interface } from '@ethersproject/abi';
+import { successfulFetch } from '@metamask/controller-utils';
+import type { Json } from '@metamask/utils';
 import { createModuleLogger } from '@metamask/utils';
 import { BigNumber } from 'bignumber.js';
 
 import {
-  ARBITRUM_USDC_ADDRESS,
-  CHAIN_ID_ARBITRUM,
-  CHAIN_ID_POLYGON,
+  CHAIN_ID_HYPERCORE,
+  RELAY_FALLBACK_GAS_LIMIT,
   RELAY_URL_QUOTE,
 } from './constants';
 import type { RelayQuote } from './types';
 import { TransactionPayStrategy } from '../..';
-import { NATIVE_TOKEN_ADDRESS } from '../../constants';
+import type { TransactionMeta } from '../../../../transaction-controller/src';
+import {
+  ARBITRUM_USDC_ADDRESS,
+  CHAIN_ID_ARBITRUM,
+  CHAIN_ID_POLYGON,
+  NATIVE_TOKEN_ADDRESS,
+} from '../../constants';
 import { projectLogger } from '../../logger';
 import type {
+  Amount,
   FiatValue,
   PayStrategyGetQuotesRequest,
   QuoteRequest,
   TransactionPayControllerMessenger,
   TransactionPayQuote,
 } from '../../types';
-import { calculateGasCost, calculateTransactionGasCost } from '../../utils/gas';
+import { calculateGasCost } from '../../utils/gas';
 import { getNativeToken, getTokenFiatRate } from '../../utils/token';
 
 const log = createModuleLogger(projectLogger, 'relay-strategy');
@@ -38,20 +46,15 @@ export async function getRelayQuotes(
   log('Fetching quotes', requests);
 
   try {
-    const result = requests
+    const normalizedRequests = requests
       // Ignore gas fee token requests
       .filter((r) => r.targetAmountMinimum !== '0')
       .map((r) => normalizeRequest(r));
 
-    const normalizedRequests = result.map((r) => r.request);
-    const isSkipTransaction = result.some((r) => r.isSkipTransaction);
-
-    log('Normalized requests', { normalizedRequests, isSkipTransaction });
+    log('Normalized requests', normalizedRequests);
 
     return await Promise.all(
-      normalizedRequests.map((r) =>
-        getSingleQuote(r, isSkipTransaction, request),
-      ),
+      normalizedRequests.map((r) => getSingleQuote(r, request)),
     );
   } catch (error) {
     log('Error fetching quotes', { error });
@@ -63,16 +66,14 @@ export async function getRelayQuotes(
  * Fetches a single Relay quote.
  *
  * @param request  - Quote request.
- * @param isSkipTransaction - Whether to skip the transaction.
  * @param fullRequest - Full quotes request.
  * @returns  Single quote.
  */
 async function getSingleQuote(
   request: QuoteRequest,
-  isSkipTransaction: boolean,
   fullRequest: PayStrategyGetQuotesRequest,
 ): Promise<TransactionPayQuote<RelayQuote>> {
-  const { messenger } = fullRequest;
+  const { messenger, transaction } = fullRequest;
 
   try {
     const body = {
@@ -86,7 +87,11 @@ async function getSingleQuote(
       user: request.from,
     };
 
+    await processTransactions(transaction, request, body, messenger);
+
     const url = getFeatureFlags(messenger).relayQuoteUrl;
+
+    log('Request body', { body, url });
 
     const response = await successfulFetch(url, {
       method: 'POST',
@@ -95,15 +100,79 @@ async function getSingleQuote(
     });
 
     const quote = (await response.json()) as RelayQuote;
-    quote.skipTransaction = isSkipTransaction;
 
-    log('Fetched relay quote', { quote, url });
+    log('Fetched relay quote', quote);
 
     return normalizeQuote(quote, request, fullRequest);
   } catch (e) {
     log('Error fetching relay quote', e);
     throw e;
   }
+}
+
+/**
+ * Add tranasction data to request body if needed.
+ *
+ * @param transaction - Transaction metadata.
+ * @param request - Quote request.
+ * @param requestBody  - Request body to populate.
+ * @param messenger  - Controller messenger.
+ */
+async function processTransactions(
+  transaction: TransactionMeta,
+  request: QuoteRequest,
+  requestBody: Record<string, Json | undefined>,
+  messenger: TransactionPayControllerMessenger,
+) {
+  const { data, value } = transaction.txParams;
+
+  /* istanbul ignore next */
+  const hasNoParams = (!data || data === '0x') && (!value || value === '0x0');
+
+  const skipDelegation =
+    hasNoParams || request.targetChainId === CHAIN_ID_HYPERCORE;
+
+  if (skipDelegation) {
+    log('Skipping delegation as no transaction data');
+    return;
+  }
+
+  const delegation = await messenger.call(
+    'TransactionPayController:getDelegationTransaction',
+    { transaction },
+  );
+
+  const normalizedAuthorizationList = delegation.authorizationList?.map(
+    (a) => ({
+      ...a,
+      chainId: Number(a.chainId),
+      nonce: Number(a.nonce),
+      yParity: Number(a.yParity),
+    }),
+  );
+
+  const tokenTransferData = new Interface([
+    'function transfer(address to, uint256 amount)',
+  ]).encodeFunctionData('transfer', [
+    request.from,
+    request.targetAmountMinimum,
+  ]);
+
+  requestBody.authorizationList = normalizedAuthorizationList;
+  requestBody.tradeType = 'EXACT_OUTPUT';
+
+  requestBody.txs = [
+    {
+      to: request.targetTokenAddress,
+      data: tokenTransferData,
+      value: '0x0',
+    },
+    {
+      to: delegation.to,
+      data: delegation.data,
+      value: delegation.value,
+    },
+  ];
 }
 
 /**
@@ -127,7 +196,9 @@ function normalizeRequest(request: QuoteRequest) {
     sourceTokenAddress: isPolygonNativeSource
       ? NATIVE_TOKEN_ADDRESS
       : request.sourceTokenAddress,
-    targetChainId: isHyperliquidDeposit ? toHex(1337) : request.targetChainId,
+    targetChainId: isHyperliquidDeposit
+      ? CHAIN_ID_HYPERCORE
+      : request.targetChainId,
     targetTokenAddress: isHyperliquidDeposit
       ? '0x00000000000000000000000000000000'
       : request.targetTokenAddress,
@@ -143,10 +214,7 @@ function normalizeRequest(request: QuoteRequest) {
     });
   }
 
-  return {
-    request: requestOutput,
-    isSkipTransaction: isHyperliquidDeposit,
-  };
+  return requestOutput;
 }
 
 /**
@@ -162,9 +230,9 @@ function normalizeQuote(
   request: QuoteRequest,
   fullRequest: PayStrategyGetQuotesRequest,
 ): TransactionPayQuote<RelayQuote> {
-  const { messenger, transaction } = fullRequest;
+  const { messenger } = fullRequest;
   const { details } = quote;
-  const { currencyIn, currencyOut } = details;
+  const { currencyIn } = details;
 
   const { usdToFiatRate } = getFiatRates(messenger, request);
 
@@ -174,18 +242,22 @@ function normalizeQuote(
   );
 
   const provider = getFiatValueFromUsd(
-    new BigNumber(currencyIn.amountUsd).minus(currencyOut.amountUsd),
+    calculateProviderFee(quote),
     usdToFiatRate,
   );
 
   const sourceNetwork = calculateSourceNetworkCost(quote, messenger);
 
-  const targetNetwork = quote.skipTransaction
-    ? {
-        usd: '0',
-        fiat: '0',
-      }
-    : calculateTransactionGasCost(transaction, messenger);
+  const targetNetwork = {
+    usd: '0',
+    fiat: '0',
+  };
+
+  const sourceAmount: Amount = {
+    human: currencyIn.amountFormatted,
+    raw: currencyIn.amount,
+    ...getFiatValueFromUsd(new BigNumber(currencyIn.amountUsd), usdToFiatRate),
+  };
 
   return {
     dust,
@@ -197,6 +269,7 @@ function normalizeQuote(
     },
     original: quote,
     request,
+    sourceAmount,
     strategy: TransactionPayStrategy.Relay,
   };
 }
@@ -308,28 +381,67 @@ function getFeatureFlags(messenger: TransactionPayControllerMessenger) {
 function calculateSourceNetworkCost(
   quote: RelayQuote,
   messenger: TransactionPayControllerMessenger,
-) {
-  const allParams = quote.steps.flatMap((s) => s.items.map((i) => i.data));
+): TransactionPayQuote<RelayQuote>['fees']['sourceNetwork'] {
+  const allParams = quote.steps[0].items.map((i) => i.data);
+  const { chainId } = allParams[0];
+  const totalGasLimit = calculateSourceNetworkGasLimit(allParams);
 
-  const result = allParams.reduce(
-    (total, params) => {
-      const gasCost = calculateGasCost({
-        ...params,
-        maxFeePerGas: undefined,
-        maxPriorityFeePerGas: undefined,
-        messenger,
-      });
+  const estimate = calculateGasCost({
+    chainId,
+    gas: totalGasLimit,
+    messenger,
+  });
 
-      return {
-        usd: new BigNumber(total.usd).plus(gasCost.usd),
-        fiat: new BigNumber(total.fiat).plus(gasCost.fiat),
-      };
-    },
-    { usd: new BigNumber(0), fiat: new BigNumber(0) },
+  const max = calculateGasCost({
+    chainId,
+    gas: totalGasLimit,
+    messenger,
+    isMax: true,
+  });
+
+  return { estimate, max };
+}
+
+/**
+ * Calculate the total gas limit for the source network transactions.
+ *
+ * @param params - Array of transaction parameters.
+ * @returns - Total gas limit.
+ */
+function calculateSourceNetworkGasLimit(
+  params: RelayQuote['steps'][0]['items'][0]['data'][],
+): number {
+  const allParamsHasGas = params.every((p) => p.gas !== undefined);
+
+  if (allParamsHasGas) {
+    return params.reduce(
+      (total, p) => total + new BigNumber(p.gas as string).toNumber(),
+      0,
+    );
+  }
+
+  // In future, call `TransactionController:estimateGas`
+  // or `TransactionController:estimateGasBatch` based on params length.
+
+  return params.reduce(
+    (total, p) =>
+      total + new BigNumber(p.gas ?? RELAY_FALLBACK_GAS_LIMIT).toNumber(),
+    0,
+  );
+}
+
+/**
+ * Calculate the provider fee for a Relay quote.
+ *
+ * @param quote - Relay quote.
+ * @returns - Provider fee in USD.
+ */
+function calculateProviderFee(quote: RelayQuote) {
+  const relayerFee = new BigNumber(quote.fees.relayer.amountUsd);
+
+  const valueLoss = new BigNumber(quote.details.currencyIn.amountUsd).minus(
+    quote.details.currencyOut.amountUsd,
   );
 
-  return {
-    usd: result.usd.toString(10),
-    fiat: result.fiat.toString(10),
-  };
+  return relayerFee.gt(valueLoss) ? relayerFee : valueLoss;
 }